fapolicyd/Sanity/trusted-execution: converted to be compatible with upgrade testing
This change introduces a compatibility with upgrade testing which means that just specific parts of the test are executed before the system upgrade and other parts after it.

Mainly the setup part is executed only before the upgrade and the cleanup is not executed at all.

This is controlled by the IN_PLACE_UPGRADE environment variable. If not set, the test runs the same as before (regular testing with all the phases executed).