Drop mujfamodul workaround
commit f5d622db7cb6 ("updated library to the cover Fedora selinux-policy
issues") introduced creating local policy module to workaround a policy
issue. Later, in commit 74f059eb8036 ("Remove selinux policy common as
lib dependency") the necessary dependency on
library(selinux-policy/common) was remove and the code started to
generate the module unconditionally.

This policy issue is already fixed and the workaround is not necessary anymore.

    # sesearch -A -s init_t -t unconfined_t -c fifo_file -p write
    allow domain rpm_transition_domain:fifo_file { append getattr ioctl lock read write };

Signed-off-by: Petr Lautrbach <lautrbach@redhat.com>