test if a missing EOL in the fapolicyd config file causes an error
A recently filed customer case revealed that the fapolicyd service
produces the following error message when the last line of its
configuration file is missing EOL character:

  fapolicyd[...]: Skipping line 21 in /etc/fapolicyd/fapolicyd.conf: too long

The TC reproduces the situation.

In order to honor the last line of the configuration file, the fapolicyd
code was improved.

The TC covers RHEL-65625 and RHEL-124283.