Wait until fapolicyd updates the db
`fapolicyd-cli --update` is not synchronous, it just notifies
`fapolicyd` to update the db, and if `fapolicyd` is stopped before it
can do the job, the db in /var/lib/fapolicyd is not updated yet. It
would be updated on the next start of fapolicyd.

Therefore, wait until the db changed before `fapolicyd` is stopped.

Fixes:
    [  BEGIN   ] ::   Running 'fapolicyd-cli --update'
    Fapolicyd was notified
    [   PASS   ] ::   Command 'fapolicyd-cli --update' (Expected 0, got 0)
    [   PASS   ] ::   Command 'fapStop' (Expected 0, got 0)
    [  BEGIN   ] ::   Running 'fapolicyd-cli -D | grep test_file'
    [   FAIL   ] ::   Command 'fapolicyd-cli -D | grep test_file' (Expected 0, got 1)

Signed-off-by: Petr Lautrbach <lautrbach@redhat.com>